requirements.txt
asgiref==3.8.1
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.1.0
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.2
